Works pretty well!
Still need to fix some formatting and fluff.

The main thing I have been doing is changing the way I figure out if an angle is 180 degrees.  First, I was seeing is one endpoint of a range returned an angle less than 180, and the other endpoint returned an angle larger than 180, but I realized that wouldn't work.  So, I decided to check to see if 180 - the angle is less than the angle change per tick of the given timeslice.