Added methods
I realized it is much easier have the angle between the hands as a state variable in the Clock class (rather than in the ClockSolverLine and ClockSolverAngle classes).  Thus, I calculated the change in angle per tick in this class.

I also added a few helpful methods in this class to adjust for this change.

Additionally, I edited the toString() method to be more clear.