Revamped whole class
Originally, I had been calculating the angle in this class, which was very wrong.  I made sure to do that in the clock class.  Thus, I was able to delete a bunch of state variables and methods from this class.  Also, my main method was way too cluttered, so I did all the "solving" in the solver constructor, which is much more sensible.