Now correctly accounts for inverse angles
The actual code was detecting inverse angles fine, but I made the print statements more clear.  E.g. If the user is seeking to find all 30 degree angles, at 1:00 o'clock, when the angle between the hands can be interpreted as both 30 degrees and 330 degrees, my program was counting this as a 30 degree angle, but telling the user the angle was 330 degrees at this time.  I changed it so that it correctly stated that the angle was 30 degrees.