Change to total() method
Changed the method so that if the tuple is impossible, the total amount of coins it contains is Integer.MAX_VALUE.  This is so that when comparing tuples, an impossible tuple will never be selected as the best solution.  Previously, this kept happening because impossible tuples were said to have a total of 0.