Changed output
Originally I had done a bunch of fancy stuff in order to output the countries' boarders nicely, but it messed up the Map Visualizer, so I changed it.  You can look at my previous commit to see what I did.  Everything works fine except it still prints out the boarders before saying "IMPOSSIBLE MAP" if it is an impossible map.  This is really just a matter of preference.